+++ drivers/ide/buddha.c
@@ -28,7 +28,6 @@
 #include <asm/amigahw.h>
 #include <asm/amigaints.h>
 
-
     /*
      *  The Buddha has 2 IDE interfaces, the Catweasel has 3, X-Surf has 2
      */
@@ -84,7 +83,6 @@ static int xsurf_irqports[XSURF_NUM_HWIFS] __initdata = {
 };
 
 #define BUDDHA_IRQ_MR	0xfc0		/* master interrupt enable */
-
 
     /*
      *  Board information
@@ -198,7 +196,7 @@ fail_base2:
 				continue;
 			}
 		}	  
-		buddha_board = (unsigned long)ZTWO_VADDR(board);
+		buddha_board = ZTWO_VADDR(board);
 		
 		/* write to BUDDHA_IRQ_MR to enable the board IRQ */
 		/* X-Surf doesn't have this.  IRQs are always on */
